+++ sdk/appconfiguration/azure-appconfiguration-provider/azure/appconfiguration/provider/_constants.py
@@ -7,6 +7,8 @@
 FEATURE_MANAGEMENT_KEY = "FeatureManagementFeatureFlags"
 FEATURE_FLAG_PREFIX = ".appconfig.featureflag/"
 
+EMPTY_LABEL = "\0"
+
 RequestTracingDisabledEnvironmentVariable = "AZURE_APP_CONFIGURATION_TRACING_DISABLED"
 AzureFunctionEnvironmentVariable = "FUNCTIONS_EXTENSION_VERSION"
 AzureWebAppEnvironmentVariable = "WEBSITE_SITE_NAME"