+++ sdk/appconfiguration/azure-appconfiguration-provider/azure/appconfiguration/provider/aio/_azureappconfigurationproviderasync.py
@@ -4,7 +4,7 @@
 # license information.
 # -------------------------------------------------------------------------
 import json
-from typing import Any, Dict, Iterable, Mapping, Optional, overload, List, Tuple, TYPE_CHECKING
+from typing import Any, Dict, Iterable, Mapping, Optional, overload, List, Tuple, TYPE_CHECKING, Union, MutableMapping
 
 from azure.appconfiguration import FeatureFlagConfigurationSetting, SecretReferenceConfigurationSetting
 from azure.appconfiguration.aio import AzureAppConfigurationClient
@@ -12,13 +12,14 @@ from azure.keyvault.secrets.aio import SecretClient
 from azure.keyvault.secrets import KeyVaultSecretIdentifier
 
 from .._models import AzureAppConfigurationKeyVaultOptions, SettingSelector
-from .._constants import FEATURE_MANAGEMENT_KEY, FEATURE_FLAG_PREFIX
+from .._constants import FEATURE_MANAGEMENT_KEY, FEATURE_FLAG_PREFIX, EMPTY_LABEL
 from .._azureappconfigurationprovider import _is_json_content_type, _get_correlation_context
 from .._user_agent import USER_AGENT
 
 if TYPE_CHECKING:
     from azure.core.credentials_async import AsyncTokenCredential
 
+JSON = MutableMapping[str, Any]  # pylint: disable=unsubscriptable-object
 
 @overload
 async def load_provider(
@@ -76,8 +77,8 @@ async def load_provider(*args, **kwargs) -> "AzureAppConfigurationProvider":
     credential: Optional["AsyncTokenCredential"] = kwargs.pop("credential", None)
     connection_string: Optional[str] = kwargs.pop("connection_string", None)
     key_vault_options: Optional[AzureAppConfigurationKeyVaultOptions] = kwargs.pop("key_vault_options", None)
-    selects: List[SettingSelector] = kwargs.pop("selects", [SettingSelector("*", "\0")])
-    trim_prefixes : List[str] = kwargs.pop("trimmed_key_prefixes", [])
+    selects: List[SettingSelector] = kwargs.pop("selects", [SettingSelector(key_filter="*", label_filter=EMPTY_LABEL)])
+    trim_prefixes : List[str] = kwargs.pop("trim_prefixes", [])
 
     # Update endpoint and credential if specified positionally.
     if len(args) > 2:
@@ -100,10 +101,6 @@ async def load_provider(*args, **kwargs) -> "AzureAppConfigurationProvider":
     provider = _buildprovider(connection_string, endpoint, credential, key_vault_options)
 
     provider._trim_prefixes = sorted(trim_prefixes, key=len, reverse=True)
-
-    if key_vault_options is not None and len(key_vault_options.secret_clients) > 0:
-        for secret_client in key_vault_options.secret_clients:
-            provider._secret_clients[secret_client.vault_url] = secret_client
 
     for select in selects:
         configurations = provider._client.list_configuration_settings(
@@ -183,9 +180,16 @@ async def _resolve_keyvault_reference(
 
     if referenced_client is None and key_vault_options.credential is not None:
         referenced_client = SecretClient(
-            vault_url=key_vault_identifier.vault_url, credential=key_vault_options.credential
+            vault_url=key_vault_identifier.vault_url, credential=key_vault_options.credential, 
+             **key_vault_options.client_configs.get(key_vault_identifier.vault_url, {})
         )
         provider._secret_clients[key_vault_identifier.vault_url] = referenced_client
+    elif referenced_client is None:
+        options = key_vault_options.client_configs[key_vault_identifier.vault_url]
+        referenced_client = SecretClient(
+            vault_url=key_vault_identifier.vault_url, credential=options.pop("credential"), 
+             **options
+        )
 
     if referenced_client:
         return (
@@ -206,7 +210,7 @@ async def _resolve_keyvault_reference(
     )
 
 
-class AzureAppConfigurationProvider(Mapping[str, str]):
+class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):
     """
     Provides a dictionary-like interface to Azure App Configuration settings. Enables loading of sets of configuration
     settings from Azure App Configuration into a Python application. Enables trimming of prefixes from configuration