+++ sdk/appconfiguration/azure-appconfiguration-provider/samples/async_key_vault_reference_provided_clients_sample.py
@@ -20,7 +20,7 @@ async def main():
 
     # Connection to Azure App Configuration using AAD with Provided Client
     secret_client = SecretClient(vault_url=key_vault_uri, credential=credential)
-    selects = {SettingSelector("*", "prod")}
+    selects = {SettingSelector(key_filter="*", label_filter="prod")}
     key_vault_options = AzureAppConfigurationKeyVaultOptions(secret_clients=[secret_client])
     config = await load_provider(endpoint=endpoint, credential=credential, key_vault_options=key_vault_options, selects=selects)