+++ sdk/appconfiguration/azure-appconfiguration-provider/samples/async_key_vault_reference_sample.py
@@ -18,7 +18,7 @@ async def main():
 
     # Connection to Azure App Configuration using AAD and Resolving Key Vault References
     key_vault_options = AzureAppConfigurationKeyVaultOptions(credential=credential)
-    selects = {SettingSelector("*", "prod")}
+    selects = {SettingSelector(key_filter="*", label_filter="prod")}
 
     config = await load_provider(endpoint=endpoint, credential=credential, key_vault_options=key_vault_options, selects=selects)