+++ sdk/appconfiguration/azure-appconfiguration-provider/samples/connection_string_sample.py
@@ -25,7 +25,7 @@ config = load_provider(connection_string=connection_string, trimmed_key_prefixes
 print(config["message"])
 
 # Connection to Azure App Configuration using SettingSelector
-selects = {SettingSelector("message*", "\0")}
+selects = {SettingSelector(key_filter="message*")}
 config = load_provider(connection_string=connection_string, selects=selects)
 
 print("message found: " + str("message" in config))