+++ sdk/appconfiguration/azure-appconfiguration-provider/samples/key_vault_reference_customized_clients_sample.py
@@ -9,7 +9,6 @@ from azure.appconfiguration.provider import (
     AzureAppConfigurationKeyVaultOptions,
     SettingSelector
 )
-from azure.keyvault.secrets import SecretClient
 import os
 from sample_utilities import get_authority, get_audience, get_credential
 
@@ -20,9 +19,9 @@ audience = get_audience(authority)
 credential = get_credential(authority)
 
 # Connection to Azure App Configuration using AAD with Provided Client
-secret_client = SecretClient(vault_url=key_vault_uri, credential=credential)
-selects = {SettingSelector("*", "prod")}
-key_vault_options = AzureAppConfigurationKeyVaultOptions(secret_clients=[secret_client])
+client_configs = {key_vault_uri: {'credential': credential}}
+selects = {SettingSelector(key_filter="*", label_filter="prod")}
+key_vault_options = AzureAppConfigurationKeyVaultOptions(client_configs=client_configs)
 config = load_provider(endpoint=endpoint, credential=credential, key_vault_options=key_vault_options, selects=selects)
 
 print(config["secret"])