+++ sdk/appconfiguration/azure-appconfiguration-provider/samples/key_vault_reference_sample.py
@@ -19,7 +19,7 @@ credential = get_credential(authority)
 
 # Connection to Azure App Configuration using AAD and Resolving Key Vault References
 key_vault_options = AzureAppConfigurationKeyVaultOptions(credential=credential)
-selects = {SettingSelector("*", "prod")}
+selects = {SettingSelector(key_filter="*", label_filter="prod")}
 
 config = load_provider(endpoint=endpoint, credential=credential, key_vault_options=key_vault_options, selects=selects)