+++ sdk/appconfiguration/azure-appconfiguration-provider/tests/test_async_provider_aad.py
@@ -12,9 +12,9 @@ from async_preparers import app_config_decorator_async
 
 class TestAppConfigurationProvider(AzureRecordedTestCase):
 
-    async def build_provider_aad(self, endpoint, trimmed_key_prefixes=[], selects={SettingSelector("*", "\0")}):
+    async def build_provider_aad(self, endpoint, trim_prefixes=[], selects={SettingSelector("*", "\0")}):
         cred = self.get_credential(AzureAppConfigurationClient, is_async=True)
-        return await load_provider(credential=cred, endpoint=endpoint, trimmed_key_prefixes=trimmed_key_prefixes, selects=selects)
+        return await load_provider(credential=cred, endpoint=endpoint, trim_prefixes=trim_prefixes, selects=selects)
 
     # method: provider_creation_aad
     @app_config_decorator_async
@@ -25,12 +25,12 @@ class TestAppConfigurationProvider(AzureRecordedTestCase):
             assert client["my_json"]["key"] == "value"
             assert client["FeatureManagementFeatureFlags"]["Alpha"] == '{\"enabled\": false, \"conditions\": {\"client_filters\": []}}'
 
-    # method: provider_trimmed_key_prefixes
+    # method: provider_ttrim_prefixes
     @app_config_decorator_async
     @recorded_by_proxy_async
-    async def test_provider_trimmed_key_prefixes(self, appconfiguration_endpoint_string):
+    async def test_provider_trim_prefixes(self, appconfiguration_endpoint_string):
         trimmed = {"test."}
-        async with await self.build_provider_aad(appconfiguration_endpoint_string, trimmed_key_prefixes=trimmed) as client:
+        async with await self.build_provider_aad(appconfiguration_endpoint_string, trim_prefixes=trimmed) as client:
             assert client["message"] == "hi"
             assert client["my_json"]["key"] == "value"
             assert client["trimmed"] == "key"
@@ -41,7 +41,7 @@ class TestAppConfigurationProvider(AzureRecordedTestCase):
     @app_config_decorator_async
     @recorded_by_proxy_async
     async def test_provider_selectors(self, appconfiguration_endpoint_string):
-        selects = {SettingSelector("message*", "dev")}
+        selects = {SettingSelector(key_filter="message*",label_filter="dev")}
         async with await self.build_provider_aad(appconfiguration_endpoint_string, selects=selects) as client:
             assert client["message"] == "test"
             assert "test.trimmed" not in client