+++ sdk/appconfiguration/azure-appconfiguration-provider/tests/test_provider.py
@@ -11,13 +11,12 @@ from devtools_testutils import (
     AzureRecordedTestCase, 
     recorded_by_proxy
 )
-from azure.appconfiguration import AzureAppConfigurationClient
 from preparers import app_config_decorator
 
 class TestAppConfigurationProvider(AzureRecordedTestCase):
 
-    def build_provider(self, connection_string, trimmed_key_prefixes=[], selects={SettingSelector("*", "\0")}):
-        return load_provider(connection_string=connection_string, trimmed_key_prefixes=trimmed_key_prefixes, selects=selects)
+    def build_provider(self, connection_string, trim_prefixes=[], selects={SettingSelector("*", "\0")}):
+        return load_provider(connection_string=connection_string, trim_prefixes=trim_prefixes, selects=selects)
 
     # method: provider_creation
     @recorded_by_proxy
@@ -28,12 +27,12 @@ class TestAppConfigurationProvider(AzureRecordedTestCase):
         assert client["my_json"]["key"] == "value"
         assert client["FeatureManagementFeatureFlags"]["Alpha"] == '{\"enabled\": false, \"conditions\": {\"client_filters\": []}}'
 
-    # method: provider_trimmed_key_prefixes
+    # method: provider_trim_prefixes
     @recorded_by_proxy
     @app_config_decorator
-    def test_provider_trimmed_key_prefixes(self, appconfiguration_connection_string):
+    def test_provider_trim_prefixes(self, appconfiguration_connection_string):
         trimmed = {"test."}
-        client = self.build_provider(appconfiguration_connection_string, trimmed_key_prefixes=trimmed)
+        client = self.build_provider(appconfiguration_connection_string, trim_prefixes=trimmed)
         assert client["message"] == "hi"
         assert client["my_json"]["key"] == "value"
         assert client["trimmed"] == "key"
@@ -44,7 +43,7 @@ class TestAppConfigurationProvider(AzureRecordedTestCase):
     @recorded_by_proxy
     @app_config_decorator
     def test_provider_selectors(self, appconfiguration_connection_string):
-        selects = {SettingSelector("message*", "dev")}
+        selects = {SettingSelector(key_filter="message*",label_filter="dev")}
         client = self.build_provider(appconfiguration_connection_string, selects=selects)
         assert client["message"] == "test"
         assert "test.trimmed" not in client