+++ sdk/appconfiguration/azure-appconfiguration-provider/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0b3 (Unreleased)
+## 1.0.0 (2023-03-08)
 
 ### Features Added